model_registry:  Initialize model reigstry repository
Intializing the main structure of the project in order to provide
a simple design and give a taste on the componenets needed to
establish the necessary steps.

Since this is a local `git init`, no `.github` is provided yet,
so we can differentiate the pipeline later on.

Intially, let's have soemthing functional with the Dockerfile, which
builds successfully the build image, and the application in Python,
which uses Sqlite3 and Flask to ease the programming part and the
Sqlite3 for timing and efficiency. Sqlite3 uses files mostly for the
the database.

requirements.txt is a neat way to declare python dependencies,
to differentiate are actual dependencies on the project.

Within app/ lives the python application with a [POST, GET, GET]
operations at files as required.

Signed-off-by: Manolic <echatzip@proton.me>