README:  Update README file with various information
Updating the README.md filea in order to be aligned with the
HEAD of the source code.  That is, what is debugged and at which
level.
Also, while at it, provided information about the tools used, like
`gh` and the extension `act` locally. and also, what is missing,
and what needs to be done.

Signed-off-by: Manolic <echatzip@proton.me>